test: reorganize tests folders structure
