+++ test/rules/one-level/invalid-settings.spec.js
@@ -1,7 +1,7 @@
-const { ELEMENT_TYPES: RULE } = require("../../../../src/constants/rules");
-const { SETTINGS, createRuleTester, pathResolvers } = require("../../helpers");
+const { ELEMENT_TYPES: RULE } = require("../../../src/constants/rules");
+const { SETTINGS, createRuleTester, pathResolvers } = require("../../support/helpers");
 
-const rule = require(`../../../../src/rules/${RULE}`);
+const rule = require(`../../../src/rules/${RULE}`);
 
 const { absoluteFilePath } = pathResolvers("one-level");
