+++ test/rules/two-levels/element-types.spec.js
@@ -1,7 +1,7 @@
-const { ELEMENT_TYPES: RULE } = require("../../../../src/constants/rules");
-const { SETTINGS, createRuleTester, pathResolvers } = require("../../helpers");
+const { ELEMENT_TYPES: RULE } = require("../../../src/constants/rules");
+const { SETTINGS, createRuleTester, pathResolvers } = require("../../support/helpers");
 
-const rule = require(`../../../../src/rules/${RULE}`);
+const rule = require(`../../../src/rules/${RULE}`);
 
 const errorMessage = (fileType, dependencyType) =>
   `Usage of '${dependencyType}' is not allowed in '${fileType}'`;
