+++ test/rules/two-levels/external.spec.js
@@ -1,7 +1,7 @@
-const { EXTERNAL: RULE } = require("../../../../src/constants/rules");
-const { SETTINGS, createRuleTester, pathResolvers } = require("../../helpers");
+const { EXTERNAL: RULE } = require("../../../src/constants/rules");
+const { SETTINGS, createRuleTester, pathResolvers } = require("../../support/helpers");
 
-const rule = require(`../../../../src/rules/${RULE}`);
+const rule = require(`../../../src/rules/${RULE}`);
 
 const errorMessage = (elementType, dependencyName) =>
   `Usage of external module '${dependencyName}' is not allowed in '${elementType}'`;
