+++ test/rules/two-levels/no-private.spec.js
@@ -1,7 +1,7 @@
-const { NO_PRIVATE: RULE } = require("../../../../src/constants/rules");
-const { SETTINGS, createRuleTester, pathResolvers } = require("../../helpers");
+const { NO_PRIVATE: RULE } = require("../../../src/constants/rules");
+const { SETTINGS, createRuleTester, pathResolvers } = require("../../support/helpers");
 
-const rule = require(`../../../../src/rules/${RULE}`);
+const rule = require(`../../../src/rules/${RULE}`);
 
 const errorMessage = () => `Dependency is private of another element`;
